BR-TAD-ATCNet/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer==3.4.6
colorama==0.4.6
cupy-cuda11x==11.6.0
cycler==0.11.0
decorator==5.1.1
fastrlock==0.8.3
flatbuffers==2.0.7
fonttools==4.38.0
gast==0.4.0
google-auth==2.45.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.6
joblib==1.3.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mne==1.3.1
numpy==1.20.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
platformdirs==4.0.0
pooch==1.8.2
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.67.3
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
